I have selected a subset of the tests from all the services that we can run with a release to make sure things are working
I have also made it so the axe modules are not loaded unless we are running accessilbility tests.
This prevents an issue where we cannot publish reports when the axe matchers are loaded.

Finally, I have made sure the smoulder tests (and all the tests for that matter) run in production